+++ kroxylicious-filters/kroxylicious-record-validation/src/test/java/io/kroxylicious/proxy/filter/validation/config/ValidationConfigTest.java
@@ -8,19 +8,26 @@ package io.kroxylicious.proxy.filter.validation.config;
 
 import java.net.MalformedURLException;
 import java.net.URI;
+import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 import java.util.Set;
 
+import org.jose4j.jws.AlgorithmIdentifiers;
 import org.junit.jupiter.api.Test;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 
+import io.kroxylicious.proxy.config.tls.AllowDeny;
+
+import static io.kroxylicious.test.jws.JwsTestUtils.ECDSA_VERIFY_JWKS;
+import static io.kroxylicious.test.jws.JwsTestUtils.RSA_AND_ECDSA_VERIFY_JWKS;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class ValidationConfigTest {
-
     @Test
     void testDecodeDefaultValues() throws JsonProcessingException {
         ObjectMapper yamlMapper = new ObjectMapper(new YAMLFactory());
@@ -38,10 +45,10 @@ class ValidationConfigTest {
                 """);
 
         TopicMatchingRecordValidationRule ruleOne = new TopicMatchingRecordValidationRule(Set.of("one"), null,
-                new BytebufValidation(new SyntacticallyCorrectJsonConfig(false), null, true, false));
-        TopicMatchingRecordValidationRule ruleTwo = new TopicMatchingRecordValidationRule(Set.of("two"), new BytebufValidation(null, null, true, false), null);
+                new BytebufValidation(new SyntacticallyCorrectJsonConfig(false), null, null, true, false));
+        TopicMatchingRecordValidationRule ruleTwo = new TopicMatchingRecordValidationRule(Set.of("two"), new BytebufValidation(null, null, null, true, false), null);
         ValidationConfig expected = new ValidationConfig(List.of(ruleOne, ruleTwo),
-                new RecordValidationRule(null, new BytebufValidation(null, null, true, false)));
+                new RecordValidationRule(null, new BytebufValidation(null, null, null, true, false)));
         assertEquals(expected, deserialised);
     }
 
@@ -69,9 +76,9 @@ class ValidationConfigTest {
                 """);
 
         TopicMatchingRecordValidationRule ruleOne = new TopicMatchingRecordValidationRule(Set.of("one"), null,
-                new BytebufValidation(new SyntacticallyCorrectJsonConfig(true), null, false, true));
-        TopicMatchingRecordValidationRule ruleTwo = new TopicMatchingRecordValidationRule(Set.of("two"), new BytebufValidation(null, null, false, true), null);
-        ValidationConfig expected = new ValidationConfig(List.of(ruleOne, ruleTwo), new RecordValidationRule(null, new BytebufValidation(null, null, false, true)));
+                new BytebufValidation(new SyntacticallyCorrectJsonConfig(true), null, null, false, true));
+        TopicMatchingRecordValidationRule ruleTwo = new TopicMatchingRecordValidationRule(Set.of("two"), new BytebufValidation(null, null, null, false, true), null);
+        ValidationConfig expected = new ValidationConfig(List.of(ruleOne, ruleTwo), new RecordValidationRule(null, new BytebufValidation(null, null, null, false, true)));
         assertEquals(expected, deserialised);
     }
 
@@ -92,10 +99,10 @@ class ValidationConfigTest {
                 """);
 
         TopicMatchingRecordValidationRule ruleOne = new TopicMatchingRecordValidationRule(Set.of("one"), null,
-                new BytebufValidation(new SyntacticallyCorrectJsonConfig(false), null, true, false));
-        TopicMatchingRecordValidationRule ruleTwo = new TopicMatchingRecordValidationRule(Set.of("two"), new BytebufValidation(null, null, true, false), null);
+                new BytebufValidation(new SyntacticallyCorrectJsonConfig(false), null, null, true, false));
+        TopicMatchingRecordValidationRule ruleTwo = new TopicMatchingRecordValidationRule(Set.of("two"), new BytebufValidation(null, null, null, true, false), null);
         ValidationConfig expected = new ValidationConfig(List.of(ruleOne, ruleTwo),
-                new RecordValidationRule(null, new BytebufValidation(null, null, true, false)));
+                new RecordValidationRule(null, new BytebufValidation(null, null, null, true, false)));
         assertEquals(expected, deserialised);
     }
 
@@ -126,10 +133,10 @@ class ValidationConfigTest {
                 """);
 
         TopicMatchingRecordValidationRule ruleOne = new TopicMatchingRecordValidationRule(Set.of("one"), null,
-                new BytebufValidation(new SyntacticallyCorrectJsonConfig(true), new SchemaValidationConfig(URI.create("http://localhost:8080").toURL(), 1L), false,
+                new BytebufValidation(new SyntacticallyCorrectJsonConfig(true), new SchemaValidationConfig(URI.create("http://localhost:8080").toURL(), 1L), null, false,
                         true));
-        TopicMatchingRecordValidationRule ruleTwo = new TopicMatchingRecordValidationRule(Set.of("two"), new BytebufValidation(null, null, false, true), null);
-        ValidationConfig expected = new ValidationConfig(List.of(ruleOne, ruleTwo), new RecordValidationRule(null, new BytebufValidation(null, null, false, true)));
+        TopicMatchingRecordValidationRule ruleTwo = new TopicMatchingRecordValidationRule(Set.of("two"), new BytebufValidation(null, null, null, false, true), null);
+        ValidationConfig expected = new ValidationConfig(List.of(ruleOne, ruleTwo), new RecordValidationRule(null, new BytebufValidation(null, null, null, false, true)));
         assertEquals(expected, deserialised);
     }
 
@@ -157,9 +164,96 @@ class ValidationConfigTest {
                 """);
 
         TopicMatchingRecordValidationRule ruleOne = new TopicMatchingRecordValidationRule(Set.of("one"), null,
-                new BytebufValidation(new SyntacticallyCorrectJsonConfig(true), null, false, true));
-        TopicMatchingRecordValidationRule ruleTwo = new TopicMatchingRecordValidationRule(Set.of("two"), new BytebufValidation(null, null, false, true), null);
-        ValidationConfig expected = new ValidationConfig(List.of(ruleOne, ruleTwo), new RecordValidationRule(null, new BytebufValidation(null, null, false, true)));
+                new BytebufValidation(new SyntacticallyCorrectJsonConfig(true), null, null, false, true));
+        TopicMatchingRecordValidationRule ruleTwo = new TopicMatchingRecordValidationRule(Set.of("two"), new BytebufValidation(null, null, null, false, true), null);
+        ValidationConfig expected = new ValidationConfig(List.of(ruleOne, ruleTwo), new RecordValidationRule(null, new BytebufValidation(null, null, null, false, true)));
         assertEquals(expected, deserialised);
+    }
+
+    @Test
+    void testDecodeDefaultValuesJwsSignatureValidation() throws JsonProcessingException {
+        ObjectMapper yamlMapper = new ObjectMapper(new YAMLFactory());
+        ValidationConfig deserialised = yamlMapper.readerFor(ValidationConfig.class).readValue("""
+                defaultRule:
+                  valueRule: {}
+                rules:
+                - topicNames:
+                  - one
+                  valueRule:
+                    syntacticallyCorrectJson: {}
+                    jwsSignatureValidation:
+                        trustedJsonWebKeySet: >
+                            %s
+                - topicNames:
+                  - two
+                  keyRule: {}
+                """.formatted(ECDSA_VERIFY_JWKS.toJson()));
+
+        TopicMatchingRecordValidationRule ruleOne = new TopicMatchingRecordValidationRule(Set.of("one"), null,
+                new BytebufValidation(new SyntacticallyCorrectJsonConfig(false), null, new JwsSignatureValidationConfig(ECDSA_VERIFY_JWKS, null, null, false, true), true,
+                        false));
+        TopicMatchingRecordValidationRule ruleTwo = new TopicMatchingRecordValidationRule(Set.of("two"), new BytebufValidation(null, null, null, true, false), null);
+        ValidationConfig expected = new ValidationConfig(List.of(ruleOne, ruleTwo),
+                new RecordValidationRule(null, new BytebufValidation(null, null, null, true, false)));
+        assertEquals(expected, deserialised);
+        assertEquals(getJwsConfigHashCode(expected), getJwsConfigHashCode(deserialised));
+    }
+
+    @Test
+    void testDecodeNonDefaultValuesJwsSignatureValidation() throws JsonProcessingException {
+        ObjectMapper yamlMapper = new ObjectMapper(new YAMLFactory());
+        ValidationConfig deserialised = yamlMapper.readerFor(ValidationConfig.class).readValue(
+                """
+                        defaultRule:
+                          valueRule:
+                            allowNulls: false
+                            allowEmpty: true
+                        rules:
+                        - topicNames:
+                          - one
+                          valueRule:
+                            syntacticallyCorrectJson:
+                                validateObjectKeysUnique: true
+                            jwsSignatureValidation:
+                                trustedJsonWebKeySet: >
+                                    %s
+                                algorithms:
+                                    allowed:
+                                        - ES256
+                                        - RS256
+                                jwsRecordHeaderKey: kroxylicious.io/jws
+                                contentDetached: true
+                                failOnMissingJwsRecordHeader: false
+                            allowNulls: false
+                            allowEmpty: true
+                        - topicNames:
+                          - two
+                          keyRule:
+                            allowNulls: false
+                            allowEmpty: true
+                        """.formatted(RSA_AND_ECDSA_VERIFY_JWKS.toJson()));
+
+        TopicMatchingRecordValidationRule ruleOne = new TopicMatchingRecordValidationRule(Set.of("one"), null,
+                new BytebufValidation(new SyntacticallyCorrectJsonConfig(true), null,
+                        new JwsSignatureValidationConfig(RSA_AND_ECDSA_VERIFY_JWKS,
+                                new AllowDeny<>(List.of(AlgorithmIdentifiers.ECDSA_USING_P256_CURVE_AND_SHA256, AlgorithmIdentifiers.RSA_USING_SHA256),
+                                        Collections.emptySet()),
+                                "kroxylicious.io/jws", true, false),
+                        false,
+                        true));
+        TopicMatchingRecordValidationRule ruleTwo = new TopicMatchingRecordValidationRule(Set.of("two"), new BytebufValidation(null, null, null, false, true), null);
+        ValidationConfig expected = new ValidationConfig(List.of(ruleOne, ruleTwo), new RecordValidationRule(null, new BytebufValidation(null, null, null, false, true)));
+        assertEquals(expected, deserialised);
+        assertEquals(getJwsConfigHashCode(expected), getJwsConfigHashCode(deserialised));
+    }
+
+    private static int getJwsConfigHashCode(ValidationConfig config) {
+        Optional<BytebufValidation> valueRule = Objects.requireNonNull(config.rules()).get(0).getValueRule();
+
+        if (valueRule.isEmpty()) {
+            throw new IllegalArgumentException("No value rule found in ValidationConfig, cannot get hashCode for getJwsSignatureValidationConfig");
+        }
+
+        return valueRule.get().getJwsSignatureValidationConfig().hashCode();
     }
 }