+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Format `<github issue/pr number>: <short description>`.
 * [#2922](https://github.com/kroxylicious/kroxylicious/pull/2922): build(deps): bump kafka.version from 4.1.0 to 4.1.1
 * [#1318](https://github.com/kroxylicious/kroxylicious/issues/1318): Add FilterContext#topicNames to enable filters to retrieve names for topic ids
 * [#2821](https://github.com/kroxylicious/kroxylicious/pull/2821): Fix OauthBearerValidationFilter unnecessarily copying the authentication bytes from an incoming request to a failed response 
+* [#2861](https://github.com/kroxylicious/kroxylicious/pull/2861): Add JWS Signature validator
 
 ## 0.17.1